+++ cli/src/pkg/appctx/app_config.go
@@ -106,6 +106,7 @@ type AppConfig struct {
 	KeepAlive         bool   `toml:"keep-alive,omitempty"          envconfig:"CB_KEEP_ALIVE" default:"false"`
 	SilenceBuild      bool   `toml:"silence-build,omitempty"       envconfig:"CB_SILENCE_BUILD" default:"false"`
 	Daemon            bool   `toml:"daemon,omitempty"              envconfig:"CB_DAEMON" default:"false"`
+	Browser           bool   `toml:"browser,omitempty"             envconfig:"CB_BROWSER" default:"true"`
 	Pull              bool   `toml:"pull,omitempty"                envconfig:"CB_PULL" default:"false"`
 	Dind              bool   `toml:"dind,omitempty"                envconfig:"CB_DIND" default:"false"`
 	Sudo              bool   `toml:"sudo,omitempty"                envconfig:"CB_SUDO" default:"true"`
@@ -218,6 +219,7 @@ func (config AppConfig) String() string {
 	fmt.Fprintf(&str, "    KeepAlive:         %t\n", config.KeepAlive)
 	fmt.Fprintf(&str, "    SilenceBuild:      %t\n", config.SilenceBuild)
 	fmt.Fprintf(&str, "    Daemon:            %t\n", config.Daemon)
+	fmt.Fprintf(&str, "    Browser:           %t\n", config.Browser)
 	fmt.Fprintf(&str, "    Pull:              %t\n", config.Pull)
 	fmt.Fprintf(&str, "    Dind:              %t\n", config.Dind)
 	fmt.Fprintf(&str, "    Sudo:              %t\n", config.Sudo)